Modularize filter tool
Before this commit, the filter tool was integrated in the main
application. However, in favour for modularity, the filter widget should
be a class itself and the main application should instantiate an object
of this class and put that in the corresponding design tab. So this
commit addresses the necessary steps to do that...

BUG: A segfault arises when the wires are put in the schematic. Pending
to solve that...