Added system to display an arbitrary number of graphs
In this commit there have been included a number of modifications
targeted to let the display window to plot an arbitrary number of
graphs, not just S11, S21, S22. This is still WIP since there's needed
some flexibility in order to let the user to decide what plots he/she
wants to see.

The reason behind this commit is to display the isolation parameters of
the power combiners. For the moment, I think this is ok