Added wrapper functions to shrink the schematic code
ComponentInfo, WireInfo and NodeInfo structures were converted into
classes so as to let their constructor to behave as a wrapper of the old
code. For the moment, the class variables will be public, but the idea
is turn them into private.

Moreover, the Qucs netlist is obtained from the vector of components
since this way the number of lines can be reduced (a lot). However, so
far this is done only in the power combining design class